ADD: 시세그래프 레이아웃 작업
