+++ package.json
@@ -3,13 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@nivo/core": "^0.79.0",
-    "@nivo/line": "^0.79.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
     "moment": "^2.29.3",
-    "nivo": "^0.31.0",
     "react": "^18.1.0",
     "react-daum-postcode": "^3.0.3",
     "react-dom": "^18.1.0",
@@ -18,6 +15,7 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
+    "recharts": "^2.1.10",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.3.4",
